% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBioExplorer.R
\name{CBioExplorer}
\alias{CBioExplorer}
\title{CBioExplorer: a web and standalone application for screening, validation, and annotation of cancer survival related biomarkers from molecular level to clinical settings}
\usage{
CBioExplorer(run = TRUE)
}
\description{
CBioExplorer: a web and standalone application for screening, validation, and annotation of cancer survival related biomarkers from molecular level to clinical settings
}
