% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBioProfiler.R
\name{CBioProfiler}
\alias{CBioProfiler}
\title{CBioProfiler: a web and standalone pipeine for cancer biomarker and subtype characteration from molecular level to clinical settings}
\usage{
CBioProfiler(run = TRUE)
}
\arguments{
\item{run}{}
}
\description{
CBioProfiler: a web and standalone pipeine for cancer biomarker and subtype characteration from molecular level to clinical settings
}
